рабочий фундамент

+++ database/migrations/2018_09_30_101942_create_platens_table.php
@@ -16,7 +16,7 @@ class CreatePlatensTable extends Migration
         Schema::create('platens', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title');
-            $table->unsignedSmallInteger('capacity');
+            $table->unsignedTinyInteger('capacity');
             $table->boolean('is_delete')->default(false);
             $table->timestamps();
         });
